Adding Address service tests and contact service tests

+++ src/main/java/za/ac/cput/util/Helper.java
@@ -8,7 +8,7 @@ public class Helper {
         return false;
     }
 
-    public static boolean intIsNull(long i) {
+    public static boolean intIsNull(int i) {
         if (i <= 0) {
             return true;
         }
